+++ Runtime/Patterns/Mediator/Implementation/RequestHandlerBase.cs
@@ -1,0 +1,7 @@
+namespace Patterns.Mediator.Implementation
+{
+    public abstract class RequestHandlerBase
+    {
+        public abstract object Handle(object request, object handler);
+    }
+}