+++ Runtime/Patterns/Mediator/Implementation/Unit.cs
@@ -1,0 +1,7 @@
+namespace Patterns.Mediator.Implementation
+{
+    public struct Unit
+    {
+        public static Unit Empty => new();
+    }
+}