+++ Runtime/Patterns/Mediator/Sender/IRequest.cs
@@ -1,7 +1,10 @@
 namespace Patterns.Mediator.Sender
 {
-    // ReSharper disable once UnusedTypeParameter
     public interface IRequest<T>
+    {
+    }
+
+    public interface IRequest
     {
     }
 }