+++ Runtime/Patterns/Mediator/Sender/ISender.cs
@@ -3,5 +3,6 @@ namespace Patterns.Mediator.Sender
     public interface ISender
     {
         T Send<T>(IRequest<T> request);
+        void Send<T>(T request) where T : IRequest;
     }
 }